+++ src/gdrdrv/gdrdrv.c
@@ -74,6 +74,10 @@ static int gdrdrv_cpu_must_use_device_mapping = 0;
 
 //-----------------------------------------------------------------------------
 
+static atomic64_t gdrdrv_nv_get_pages_refcount = ATOMIC_INIT(0);
+
+//-----------------------------------------------------------------------------
+
 #if LINUX_VERSION_CODE <= KERNEL_VERSION(2,6,32)
 /**
  * This API is available after Linux kernel 2.6.32
@@ -504,6 +508,10 @@ static void gdr_free_mr_unlocked(gdr_mr_t *mr)
             gdr_err("nvidia_p2p_put_pages error %d, async callback may have been fired\n", status);
         }
         #endif
+
+        if (!status) {
+            atomic64_fetch_dec(&gdrdrv_nv_get_pages_refcount);
+        }
     } else {
         gdr_dbg("invoking unpin_buffer while callback has already been fired\n");
 
@@ -531,11 +539,6 @@ static int gdrdrv_release(struct inode *inode, struct file *filp)
     if (!info) {
         gdr_err("filp contains no info\n");
         return -EIO;
-    }
-    // Check that the caller is the same process that did gdrdrv_open
-    if (!gdrdrv_check_same_process(info, current)) {
-        gdr_dbg("filp is not opened by the current process\n");
-        return -EACCES;
     }
 
     mutex_lock(&info->lock);
@@ -647,6 +650,8 @@ static void gdrdrv_get_pages_free_callback(void *data)
     page_table = mr->page_table;
     if (page_table) {
         nvidia_p2p_free_page_table(page_table);
+        // NVIDIA driver takes back those pages. Decrement the refcount here.
+        atomic64_fetch_dec(&gdrdrv_nv_get_pages_refcount);
         if (gdr_mr_is_mapped(mr))
             gdr_mr_destroy_all_mappings(mr);
     } else {
@@ -792,6 +797,10 @@ static int __gdrdrv_pin_buffer(gdr_info_t *info, u64 addr, u64 size, u64 p2p_tok
     gdr_info("invoking nvidia_p2p_get_pages(va=0x%llx len=%lld p2p_tok=%llx va_tok=%x callback=%px)\n",
              mr->va, mr->mapped_size, mr->p2p_token, mr->va_space, free_callback_fn);
     #endif
+
+    if (!ret) {
+        atomic64_fetch_inc(&gdrdrv_nv_get_pages_refcount);
+    }
 
     #ifndef CONFIG_ARM64
     tb = get_cycles();
@@ -1559,11 +1568,34 @@ static int gdrdrv_proc_params_release(struct inode *inode, struct file *filp)
     return single_release(inode, filp);
 }
 
-static const struct proc_ops gdrdrv_proc_params_pro_ops = {
+static const struct proc_ops gdrdrv_proc_params_proc_ops = {
     .proc_open = gdrdrv_proc_params_open,
     .proc_read = seq_read,
     .proc_lseek = seq_lseek,
     .proc_release = gdrdrv_proc_params_release
+};
+
+static int gdrdrv_proc_ngp_refcount_read(struct seq_file *s, void *v)
+{
+    seq_printf(s, "%lld\n", atomic64_read(&gdrdrv_nv_get_pages_refcount));
+    return 0;
+}
+
+static int gdrdrv_proc_ngp_refcount_open(struct inode *inode, struct file *filp)
+{
+    return single_open(filp, gdrdrv_proc_ngp_refcount_read, NULL);
+}
+
+static int gdrdrv_proc_ngp_refcount_release(struct inode *inode, struct file *filp)
+{
+    return single_release(inode, filp);
+}
+
+static const struct proc_ops gdrdrv_proc_ngp_refcount_proc_ops = {
+    .proc_open = gdrdrv_proc_ngp_refcount_open,
+    .proc_read = seq_read,
+    .proc_lseek = seq_lseek,
+    .proc_release = gdrdrv_proc_ngp_refcount_release
 };
 
 static int gdrdrv_procfs_init(void)
@@ -1584,10 +1616,19 @@ static int gdrdrv_procfs_init(void)
         goto out;
     }
 
-    entry = proc_create("params", entry_mode, gdrdrv_proc_dir_entry, &gdrdrv_proc_params_pro_ops);
+    entry = proc_create("params", entry_mode, gdrdrv_proc_dir_entry, &gdrdrv_proc_params_proc_ops);
     if (!entry) {
         status = EINVAL;
         goto out;
+    }
+
+    if (dbg_enabled) {
+        entry_mode = (S_IFREG | S_IRUSR | S_IRGRP);
+        entry = proc_create("nv_get_pages_refcount", entry_mode, gdrdrv_proc_dir_entry, &gdrdrv_proc_ngp_refcount_proc_ops);
+        if (!entry) {
+            status = EINVAL;
+            goto out;
+        }
     }
 
 out:
@@ -1689,12 +1730,19 @@ static int __init gdrdrv_init(void)
 
 static void __exit gdrdrv_cleanup(void)
 {
+    int64_t last_nv_get_pages_refcount;
     gdr_msg(KERN_INFO, "unregistering major number %d\n", gdrdrv_major);
 
     gdrdrv_procfs_cleanup();
 
     /* cleanup_module is never called if registering failed */
     unregister_chrdev(gdrdrv_major, DEVNAME);
+
+    last_nv_get_pages_refcount = atomic64_read(&gdrdrv_nv_get_pages_refcount);
+    if (dbg_enabled)
+        WARN_ON(0 != last_nv_get_pages_refcount);
+    else
+        BUG_ON(0 != last_nv_get_pages_refcount);
 }
 
 //-----------------------------------------------------------------------------