+++ CHANGELOG.md
@@ -6,8 +6,10 @@
 - Introduce gdr\_get\_attribute API and GDR\_ATTR\_USE\_PERSISTENT\_MAPPING.
 - Add persistent mapping support in gdrcopy\_sanity.
 - Support setting GPU ID in gdrcopy\_sanity.
+- Fix nvidia\_p2p\_get\_pages\_persistent leakage in gdrdrv.
+- Introduce leakage\_pin\_pages\_fork\_cumemalloc unit test.
 - Fix the use-after-free bug of mr objects in gdrdv\_vma\_close.
-- Introduce /proc/driver/gdrdrv/params.
+- Introduce /proc/driver/gdrdrv/params and /proc/driver/gdrdrv/nv\_get\_pages\_refcount.
 
 ## [2.4.3] - 2024-12-02
 - Fix NVIDIA\_IS\_OPENSOURCE detection when compile with NVIDIA driver version 545 or newer.